refactor(code-review): extract repo slice primitives
Move extractContentWindow, readSnippetWindow, extractModifiedFunctionNames
and generateCallGraphGrep out of call-graph.helper.ts into repo-slices.ts,
verbatim, and import them back. They were module-private and served only the
generalist review path; the Kody Rules retriever needs the same three moves.

call-graph.helper.spec.ts is unchanged and passes as the regression guard.
